fix:  [pipe-26591]: refactor details page
