+++ apps/design-system/src/subjects/views/connectors/connector-details.tsx
@@ -3,11 +3,15 @@ import { useTranslationStore } from '@utils/viewUtils'
 import { noop } from 'lodash-es'
 
 import { InputFactory } from '@harnessio/forms'
+import { Tabs } from '@harnessio/ui/components'
 import {
   ArrayInput,
   BooleanInput,
+  ConnectorDetailsConfiguration,
   ConnectorDetailsItem,
-  ConnectorDetailsPage,
+  ConnectorDetailsLayout,
+  ConnectorDetailsReferencePage,
+  ConnectorDetailsTabsKeys,
   GroupInput,
   ListInput,
   NumberInput,
@@ -33,46 +37,64 @@ inputComponentFactory.registerComponent(new SelectInput())
 inputComponentFactory.registerComponent(new SeparatorInput())
 inputComponentFactory.registerComponent(new RadialInput())
 
-const ConnectorsDetailsPageWrapper = (): JSX.Element => (
-  <ConnectorDetailsPage
-    searchQuery=""
-    toEntity={noop}
-    toScope={noop}
-    entities={mockConnectorRefList}
-    isConnectorReferencesLoading={false}
-    setIsConnectorRefSearchQuery={noop}
-    currentPage={1}
-    totalPages={1}
-    goToPage={noop}
-    connectorDetails={
-      {
-        name: mockConnectorDetails.connector.name,
-        identifier: mockConnectorDetails.connector.identifier,
-        type: mockConnectorDetails.connector.type,
-        status: mockConnectorDetails.status.status,
-        lastTestedAt: mockConnectorDetails.status.lastTestedAt,
-        lastModifiedAt: mockConnectorDetails.lastModifiedAt,
-        spec: {
-          url: mockConnectorDetails.connector.spec.url
-        },
-        gitDetails: {
-          repoIdentifier: mockConnectorDetails.gitDetails.repoIdentifier || '',
-          branch: mockConnectorDetails.gitDetails.branch || '',
-          objectId: mockConnectorDetails.gitDetails.objectId || ''
-        },
-        lastConnectedAt: mockConnectorDetails.status.lastConnectedAt,
-        createdAt: mockConnectorDetails.createdAt,
-        icon: 'github',
-        description: mockConnectorDetails.connector.description
-      } as ConnectorDetailsItem
-    }
-    useTranslationStore={useTranslationStore}
-    onTest={noop}
-    onDelete={noop}
-    onSave={noop}
-    getConnectorDefinition={type => getHarnessConnectorDefinition(type, { autoExpandGroups: true })}
-    inputComponentFactory={inputComponentFactory}
-  />
-)
+const ConnectorsDetailsPageWrapper = (): JSX.Element => {
+  const connectorDetails = {
+    name: mockConnectorDetails.connector.name,
+    identifier: mockConnectorDetails.connector.identifier,
+    type: mockConnectorDetails.connector.type,
+    status: mockConnectorDetails.status.status,
+    lastTestedAt: mockConnectorDetails.status.lastTestedAt,
+    lastModifiedAt: mockConnectorDetails.lastModifiedAt,
+    spec: {
+      url: mockConnectorDetails.connector.spec.url
+    },
+    gitDetails: {
+      repoIdentifier: mockConnectorDetails.gitDetails.repoIdentifier || '',
+      branch: mockConnectorDetails.gitDetails.branch || '',
+      objectId: mockConnectorDetails.gitDetails.objectId || ''
+    },
+    lastConnectedAt: mockConnectorDetails.status.lastConnectedAt,
+    createdAt: mockConnectorDetails.createdAt,
+    icon: 'github',
+    description: mockConnectorDetails.connector.description
+  } as ConnectorDetailsItem
+  return (
+    <ConnectorDetailsLayout
+      connectorDetails={connectorDetails}
+      onTest={noop}
+      onDelete={noop}
+      useTranslationStore={useTranslationStore}
+    >
+      <Tabs.Content className="pt-7" value={ConnectorDetailsTabsKeys.CONFIGURATION}>
+        <ConnectorDetailsConfiguration
+          connectorDetails={connectorDetails}
+          onSave={noop}
+          inputComponentFactory={inputComponentFactory}
+          getConnectorDefinition={getHarnessConnectorDefinition}
+          useTranslationStore={useTranslationStore}
+          apiError={''}
+        />
+      </Tabs.Content>
+      <Tabs.Content className="pt-7" value={ConnectorDetailsTabsKeys.REFERENCES}>
+        <ConnectorDetailsReferencePage
+          toEntity={noop}
+          toScope={noop}
+          entities={mockConnectorRefList}
+          searchQuery={''}
+          apiConnectorRefError={undefined}
+          useTranslationStore={useTranslationStore}
+          isLoading={false}
+          setSearchQuery={noop}
+          currentPage={1}
+          totalPages={1}
+          goToPage={noop}
+        />
+      </Tabs.Content>
+      <Tabs.Content className="pt-7" value={ConnectorDetailsTabsKeys.ACTIVITY}>
+        <div>Activity History</div>
+      </Tabs.Content>
+    </ConnectorDetailsLayout>
+  )
+}
 
 export { ConnectorsDetailsPageWrapper }
