+++ packages/ui/src/views/connectors/connector-details/index.ts
@@ -1,2 +1,4 @@
-export * from '@views/connectors/connector-details/connector-details-page'
-export * from '@views/connectors/connector-details/types'
+export * from './types'
+export * from './connector-details-layout'
+export * from './connector-details-configuration'
+export * from './connector-details-references-page'
