+++ packages/ui/src/views/connectors/connector-details/types.ts
@@ -99,6 +99,14 @@ export interface ConnectorDetailsPageProps {
   apiConnectorRefError?: string
 }
 
+export interface ConnectorDetailsLayoutProps {
+  connectorDetails: ConnectorDetailsItem
+  onTest: (connectorId: string) => void
+  onDelete: (connectorId: string) => void
+  useTranslationStore: () => TranslationStore
+  children: React.ReactNode
+}
+
 export interface ConnectorDetailsReferencePageProps {
   searchQuery: string
   setSearchQuery: (query?: string) => void
