+++ packages/ui/src/views/connectors/index.ts
@@ -2,8 +2,8 @@ export * from './types'
 export * from './connector-reference/connector-reference'
 export * from './connector-header'
 export * from './connectors-list'
-export * from './connector-details'
 export * from './connector-entity-form'
 export * from './connectors-pallete-drawer'
 export * from './connector-input/connector-input'
 export * from './connector-test-connection/connector-test-connection-dialog'
+export * from './connector-details/index'
